# openapi.yaml
openapi: 3.1.0
info:
  title: Go Media Transcoder Dynamic Playback API
  version: 0.1.0
  description: |
    Dynamic HLS/DASH playback-origin API backed by direct FFmpeg libav* calls.
    The service does not spawn ffmpeg/ffprobe processes at runtime and does not
    expose static transcode-to-output endpoints.
servers:
  - url: http://localhost:8080
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /v1/playback/dash/{id}/manifest.mpd:
    get:
      summary: DASH MPD manifest
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "200":
          description: MPD manifest
          content:
            application/dash+xml:
              schema: { type: string }
...
  /v1/playback/dash/{id}:
    delete:
      summary: Delete a DASH playback session and cache
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "204":
          description: Session deleted
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    SessionID:
      name: id
      in: path
      required: true
      schema:
        type: string
...